chore: Rename api directory as it deploys functions
